c-ares-1.1.0

+++ ares_expand_name.c
@@ -24,6 +24,7 @@
 
 #include <stdlib.h>
 #include "ares.h"
+#include "ares_private.h" /* for the memdebug */
 
 static int name_length(const unsigned char *encoded, const unsigned char *abuf,
 		       int alen);
@@ -53,7 +54,7 @@ static int name_length(const unsigned char *encoded, const unsigned char *abuf,
  */
 
 int ares_expand_name(const unsigned char *encoded, const unsigned char *abuf,
-		     int alen, char **s, int *enclen)
+		     int alen, char **s, long *enclen)
 {
   int len, indir = 0;
   char *q;
